+++ CHANGELOG.md
@@ -4,12 +4,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.11.20] - 18-07-2025
+ 
+### Added
+
+### Changed
+- Added `highest severity level` to output
+- Modify `Rule` dataclass to include a severity field
 
 ## [0.11.19] - 15-05-2025
  
 ### Added
 
-### Changed
+### Changeds
 - Refactor `output_transformations.py`
 - Modify `Rule` and `ValidationSettings` dataclasses